% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptileDB_acceptedNames.R
\name{reptileDB_acceptedNames}
\alias{reptileDB_acceptedNames}
\title{Get Reptile Database accepted names}
\usage{
reptileDB_acceptedNames(version = "latest")
}
\arguments{
\item{version}{which version of Reptile Database to use. Can be \code{'latest'} or a combination of month and year, such as \code{'3_2024'}. See \code{\link{reptileDB_versions}} for available versions.}
}
\value{
Returns the list of accepted species names in the
database along with higher taxonomy.
}
\description{
Returns the full list of accepted names
}
\examples{

head(reptileDB_acceptedNames())

}
