% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptileDB_match.R
\name{reptileDB_match}
\alias{reptileDB_match}
\title{Resolve taxonomy to match the Reptile Database}
\usage{
reptileDB_match(
  taxon,
  fuzzyThresh = 1,
  alternateEndings = TRUE,
  retainSubspecies = FALSE,
  yearCutoff = 1950,
  version = "latest"
)
}
\arguments{
\item{taxon}{a single taxon name}

\item{fuzzyThresh}{the acceptable number of differences \emph{per} component (genus, sp, subsp) for fuzzy matching.
If zero, no fuzzy matching. See \code{\link[utils]{adist}}.}

\item{alternateEndings}{allow for testing of alternate latin suffixes when matching to accepted names}

\item{retainSubspecies}{If subspecies provided, should subspecies be retained in result.
This only pertains to cases where the taxon matches to a known subspecies. As synonyms are mapped
to Genus_species accepted names (with no subspecies), if a taxon matches to a synonym with subspecies,
it will still map back to a Genus_species.}

\item{yearCutoff}{When matching to synonymys, what is the earliest publication year to consider.
Can be set to NULL to be ignored}

\item{version}{which version of Reptile Database to use. Can be \code{'latest'} or a combination of month and year, such as \code{'3_2024'}. See \code{\link{reptileDB_versions}} for available versions.}
}
\value{
A dataframe with matches to either accepted taxon names or synonyms.
Authority and year pertain to the accepted name if directly matched, otherwise pertain to the synonym.
If a synonym was listed multiple times with different years under the same accepted name, then the oldest authority/year is listed.
}
\description{
This function will attempt to assign taxon names to a set of Reptile Database accepted names, using both strict and fuzzy matching, and via synonymy.
}
\details{
This function makes the following attempts (assuming all options enabled):
\itemize{
\item strict matching to accepted names
\item strict matching to accepted names, but with alternate latin suffixes
\item fuzzy matching to accepted names
\item strict matching to synonyms
\item fuzzy matching to synonyms
}

It is possible to match to both an accepted name and to the synonym of another accepted name.
}
\examples{

# exact matching to accepted name
reptileDB_match('Lachesis stenophrys')

# fuzzy matching to accepted name
reptileDB_match('Anolis saxitilis')

# exact matching to synonyms
reptileDB_match('Japalura planidorsata')

# fuzzy matching to synonyms
reptileDB_match('Amphisbaena mensa')

# matching via alternate latin suffixes
reptileDB_match('Magliophis exiguum')

# inclusion of subspecies
reptileDB_match('Phrynosoma coronatum blainvillii')

# multiple hits
reptileDB_match('Chironius flavolineatus')

# querying multiple taxa
lapply(c('Coluber_constrictor', 'Anolis saxitilis', 'Xenochrophis flavipunctatus'), reptileDB_match) 

# query a taxon that did not yet exist (described in 2022)
reptileDB_match('Abronia zongolica', version = 'latest')
reptileDB_match('Abronia zongolica', version = '1_2020')

}
