% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptileDB_synonymsFromAccepted.R
\name{reptileDB_synonymsFromAccepted}
\alias{reptileDB_synonymsFromAccepted}
\title{Get Reptile Database synonyms from accepted}
\usage{
reptileDB_synonymsFromAccepted(sp, version = "latest")
}
\arguments{
\item{sp}{genus and species (separated by underscore or space)}

\item{version}{which version of Reptile Database to use. Can be \code{'latest'} or a combination of month and year, such as \code{'3_2024'}. See \code{\link{reptileDB_versions}} for available versions.}
}
\value{
Returns a list with a dataframe of synonyms for each queried accepted name.
}
\description{
Returns the synonyms that are associated with a given accepted taxon name.
}
\examples{

reptileDB_synonymsFromAccepted(c('Phrynosoma_coronatum', 'Crotalus atrox'))

}
