% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reptileDB_versions.R
\name{reptileDB_versions}
\alias{reptileDB_versions}
\title{List available versions of Reptile Database}
\usage{
reptileDB_versions()
}
\value{
Returns a table of available versions, according to the month/year of download. If Reptile Database data have already been loaded, this table will also indicate which one is currently loaded.
}
\description{
As major updates to Reptile Database are released, they are made avaible via this R package.
Use this function to list the available versions.
}
\examples{
reptileDB_versions()
}
