% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squamata_alignment.R
\name{squamata_alignment}
\alias{squamata_alignment}
\title{Retrieve squamate sequence alignments}
\usage{
squamata_alignment(
  species = NULL,
  genus = NULL,
  family = NULL,
  subclade = NULL,
  geneRegion = NULL,
  concatenate = TRUE,
  includeSphenodon = FALSE,
  seqClass = "auto"
)
}
\arguments{
\item{species}{a vector of species names that the alignment will be subset to, Default: NULL}

\item{genus}{a vector of genus names that the alignment will be subset to, Default: NULL}

\item{family}{a vector of family names that the alignment will be subset to, Default: NULL}

\item{subclade}{a vector of major squamate subclades that the tree will be subset to, Default: NULL}

\item{geneRegion}{names of gene regions to include, see details. All gene regions included if NULL}

\item{concatenate}{if TRUE (default), then alignment is concatenated, if FALSE, then separate gene-specific alignments are returned as a list.}

\item{includeSphenodon}{Should Sphenodon_punctatus be included?, Default: FALSE}

\item{seqClass}{the alignment can be returned as class \code{DNAbin} or \code{DNAStringSet}, see details. \code{'auto'} means the function will preferentially pick \code{DNAStringSet} if available.}
}
\value{
if \code{concatenate = TRUE}, then a list with the alignment as the first item, and a table as the second item specifying how the alignment divides into the requested gene regions. If \code{concatenate = FALSE}, then a list of alignments, with one for each requested gene regions.
}
\description{
This function will return the multiple sequence alignment used in the inference of the
squamate phylogeny. This alignment can be subset by taxonomy and/or by gene region.
}
\details{
The list of major subclades that can be used for subsetting are listed under \code{\link{squamata}}.

if \code{geneRegion = NULL}, then all will be returned. Otherwise some combination of the following can be specified:
\code{'ADNP', 'AHR', 'AKAP9', 'AMEL', 'BACH1', 'BDNF', 'BHLHB2', 'BMP2', 'CAND1', 'CARD4', 'CILP', 'cmos', 
'COI', 'CXCR4', 'CYTB', 'DLL1', 'ECEL', 'ENC1', 'FSHR', 'FSTL5', 'GALR1', 'GHSR', 'GPR37', 'HLCS', 'INHIBA', 
'LRRN1', 'LZTSS1', 'MKL1', 'MLL3', 'MSH6', 'ND1', 'ND2', 'ND4', 'NGFB', 'NKTR', 'NTF-3', 'PDC', 'PNN', 'PRLR', 
'PTGER4', 'PTPN', 'R35', 'RAG1_firsthalf', 'RAG1_secondhalf', 'RAG2', 'rRNA_12S', 'rRNA_16S', 'SINCAIP', 
'SLC30A1', 'SLC8A1', 'SLC8A3', 'TRAF6', 'UBN1', 'VCPIP1', 'ZEB2', 'ZFP36L1'}

This function is set up to work with two classes for sequence data: \code{\link[ape]{DNAbin}} class from the
R package ape, and \code{seqClass = \link[Biostrings]{DNAStringSet}} from the \href{https://bioconductor.org/packages/release/bioc/html/Biostrings.html}{bioconductor Biostrings R package}.
If the latter is desired, then the R package Biostrings must be installed (available via bioconductor, not CRAN).
Setting \code{seqClass = 'auto'} means the function will preferentially pick \code{DNAStringSet} because it is
faster, and will fall back on \code{DNAbin} if the Biostrings package is not installed.
}
\examples{
\dontrun{

# get the full concatenated alignment
squamata_alignment()

# get the full concatenated alignment for Gekkota
squamata_alignment(subclade = 'Gekkota')

# get the alignment for genus Lerista, and only for gene regions ADNP and ND1
squamata_alignment(genus = 'Lerista', geneRegion = c('ADNP', 'ND1'), concatenate = FALSE)

# get the alignment for genus Lerista, and only for gene regions ADNP and ND1
## this time get the concatenated alignment
squamata_alignment(genus = 'Lerista', geneRegion = c('ADNP', 'ND1'), concatenate = TRUE)
}
}
\seealso{
\code{\link[ape]{DNAbin}}, \code{\link[Biostrings]{DNAStringSet}}
}
