% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squamata_phylogeny.R
\name{squamata_phylogeny}
\alias{squamata_phylogeny}
\title{Retrieve squamate phylogenies}
\usage{
squamata_phylogeny(
  species = NULL,
  genus = NULL,
  family = NULL,
  subclade = NULL,
  type
)
}
\arguments{
\item{species}{a vector of species names that the tree will be subset to, Default: NULL}

\item{genus}{a vector of genus names that the tree will be subset to, Default: NULL}

\item{family}{a vector of family names that the tree will be subset to, Default: NULL}

\item{subclade}{a vector of major squamate subclades that the tree will be subset to, Default: NULL}

\item{type}{choice of phylogeny to retrieve, see details}
}
\value{
an object of class \code{phylo} or \code{multiPhylo}. If \code{type = 'medianTimeTree'}, then a list is returned with the tree and a table summarizing divergence times.
}
\description{
This function will retrieve the squamate phylogeny and optionally subset it.
}
\details{
The list of major subclades that can be used for subsetting are listed under \code{\link{squamata}}.

Options for argument \code{type}:
\itemize{
\item 'ultrametric': The primary time-calibrated tree, containing 6885 species.
\item 'molecular': The primary tree with 6885 species, but with branch lengths representing molecular rates of substition.
\item 'unconstrained': An alternative version of the 6885-tip tree, where topology constraints were NOT imposed.
\item 'pseudo-posterior': A set of 100 trees, each with 6885 species, representing variability in backbone topology, divergence times.
\item 'genomic': The phylogenomic tree with 1113 species. This topology constrained relationships in the 6885-tip trees.
\item 'medianTimeTree': An ultrametric tree with 137 species, resulting the MCMCtree analysis. Here, divergence times are calculated as the median node ages across the posterior distribution of trees.
\item 'timeTreesPost': A posterior distribution of 1000 ultrametric trees, as generated by MCMCtree, with 137 tips.
}
}
\examples{
\dontrun{

# get the main 6885-tip squamate phylogeny
squamata_phylogeny(type = 'ultrametric')

# get the main tree, but subset to family Phrynosomatidae
squamata_phylogeny(type = 'ultrametric', family = 'Phrynosomatidae')

# get the main ultrametric tree subset to genus Lerista and compare it
# to the Lerista tree from an topologically unconstrained analysis
Lerista_con <- squamata_phylogeny(type = 'ultrametric', genus = 'Lerista')
Lerista_uncon <- squamata_phylogeny(type = 'unconstrained', genus = 'Lerista')
ape::comparePhylo(Lerista_con, Lerista_uncon)

# get the non-time-calibrated molecular tree for a specified set of species
spset <- c('Arrhyton_dolichura', 'Masticophis_bilineatus', 'Protobothrops_cornutus', 
'Mastigodryas_melanolomus', 'Atractus_roulei', 'Dolichophis_jugularis', 'Scincella_vandenburghi', 
'Rhampholeon_beraduccii', 'Liolaemus_azarai', 'Acanthodactylus_bedriagai', 'Boaedon_perisilvestris')
squamata_phylogeny(type = 'molecular', species = spset)

# retrieve a pseudo-posterior distribution of 100 trees, all subset to genus Anolis
squamata_phylogeny(type = 'pseudo-posterior', genus = 'Anolis')

# get the phylogenomic tree, based on SqCL, AHE and/or UCE loci
squamata_phylogeny(type = 'genomic')

# get a time-calibrated phylogeny as generated by MCMCtree, where divergence times are 
# set to the median ages across the posterior distribution of divergence times.
## note that this returns a list where the tree is the first item and a table summarizing 
## node ages as the second item. 
timetree <- squamata_phylogeny(type = 'medianTimeTree')
timetree[[1]]
head(timetree[[2]])

# get 1000 random trees from the posterior distribution of the MCMCtree dating analysis.
squamata_phylogeny(type = 'timeTreesPost')

}
}
