% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squamata_speciationRates.R
\name{squamata_speciationRates}
\alias{squamata_speciationRates}
\title{Retrieve speciation rates}
\usage{
squamata_speciationRates(
  species = NULL,
  genus = NULL,
  family = NULL,
  subclade = NULL
)
}
\arguments{
\item{species}{a vector of species names that the tree will be subset to, Default: NULL}

\item{genus}{a vector of genus names that the tree will be subset to, Default: NULL}

\item{family}{a vector of family names that the tree will be subset to, Default: NULL}

\item{subclade}{a vector of major squamate subclades that the tree will be subset to, Default: NULL}
}
\value{
a data.frame with taxonomic information and speciation rates
}
\description{
This function will retrieve species-specific speciation tip rates, including
DR, BAMM and CLaDS.
}
\details{
The list of major subclades that can be used for subsetting are listed under \code{\link{squamata}}.

For DR and CLaDS, we generated 100 fully-sampled trees via phylogenetic imputation.
The tip rates returned here are averages across those 100 trees for species that are present
in our molecular phylogeny. In other words, we do not provide tip rates for those species that were
placed via imputation. For BAMM, we modeled diversification dynamics with family-level sampling fractions.
}
\examples{
\dontrun{
# get all rates for all speciation in the tree
head(squamata_speciationRates())

# get rates for Gekkkonid geckos
head(squamata_speciationRates(family = 'Gekkonidae'))

# get rates for all snakes
head(squamata_speciationRates(subclade = 'Serpentes'))

}
}
