% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squamata_traits.R
\name{squamata_traits}
\alias{squamata_traits}
\title{Retrieve eco-morphological data}
\usage{
squamata_traits(
  species = NULL,
  genus = NULL,
  family = NULL,
  subclade = NULL,
  type = "all"
)
}
\arguments{
\item{species}{a vector of species names that the tree will be subset to, Default: NULL}

\item{genus}{a vector of genus names that the tree will be subset to, Default: NULL}

\item{family}{a vector of family names that the tree will be subset to, Default: NULL}

\item{subclade}{a vector of major squamate subclades that the tree will be subset to, Default: NULL}

\item{type}{which traits to return; defaults to \code{'all'}}
}
\value{
a data.frame with taxonomic information and trait data
}
\description{
This function will return species-level ecological, morphological,
environmental, geographical data compiled for the study of squamate macroevolution.
Trait tip rates and net innovation index are also available.
}
\details{
The list of major subclades that can be used for subsetting are listed under \code{\link{squamata}}.

Argument \code{type} can be 'all', 'morph', 'env', 'eco', 'geog', 'TR', 'innovationIndex'.

'TR' refers to tip rates of trait evolution, 'innovationIndex' refers to a measure of net
innovation of species compared to the inferred squamate ancestral state.

See this page of our website for a detailed explanation of the traits dataset.

Note that this dataset contains trait data for species that are not in the phylogeny.
The column \code{inTree} defines which taxa are included in the tree.
}
\examples{
\dontrun{
# get the full dataset with no subsetting
head(squamata_traits())

# get the full dataset for Crotalus
head(squamata_traits(genus = 'Crotalus'))
 
# return only morphological traits
head(squamata_traits(genus = 'Crotalus', type = 'morph'))

# return tip rates for all of Iguania
head(squamata_traits(subclade = 'Iguania', type = 'TR'))

}
}
